// mono-repo/packages/gantry-app/src/app/greyhound-racing/models/greyhound-racing.enum.ts
export enum RunnerReserve {
    Reserve = '(RES)',
}

export enum TypeFlagCode {
    Aus = 'AU',
    Australia = 'AUS',
    Uk = 'UK',
    Usa = 'US',
    Irish = 'IE',
}

export enum Positions {
    One = '1',
    Two = '2',
    Three = '3',
}
export enum GreyhoundMarkets {
    OddsEvensMarket = 'odds',
    InsideOutSideMarket = 'inside',
}

export enum Selections {
    selectionStatus = 'suspended',
    displayStatus = 'notdisplayed',
}

export enum MaxLimitRunners {
    NonPGRTracks = 6,
    PGRTracks = 5,
}